Jv validation setup

+++ src/main/java/com/travelapp/services/UserService.java
@@ -27,6 +27,7 @@ public class UserService{
         return userRepo.getAll();
     }
 
+
     @Transactional(readOnly=true)
     public User getById(int id) { return userRepo.findById(id); }
 
@@ -42,6 +43,9 @@ public class UserService{
 
     @Transactional
     public boolean deleteUserById(int id) {
+        if(id <= 0){
+            throw new BadRequestException();
+        }
         return userRepo.deleteById(id);
     }
